Refactor: extract shared DOM/HTML helpers into DomHelpersTrait (slice 2)
Pure, behavior-preserving move of 22 broadly-shared low-level helpers out of
HtmlTransformer into HtmlToBlocks\Support\DomHelpersTrait, which HtmlTransformer
now `use`s. No logic, signature, or output changes — trait methods are inlined
into the class so private state/method access is unchanged.

Refs #242 (slice 2: shared DOM helpers). Unblocks future slices that depend on
attr/elementSelector/safeFallbackHtml/boundedFallbackHtml/normalizedNavigationLabel
and friends without dragging HtmlTransformer along.

Parity 128->128, canonical contracts green, full `composer test` green.
